Fix bootloader signatures
