+++ skycoin-api/skycoin_constants.h
@@ -1,0 +1,9 @@
+#ifndef SKYCOIN_CONSTANTS_H
+#define SKYCOIN_CONSTANTS_H
+
+#define SKYCOIN_PUBKEY_LEN 33
+#define SKYCOIN_SECKEY_LEN 32
+#define SKYCOIN_SIG_LEN 65
+#define SKYCOIN_ADDRESS_CHECKSUM_LENGTH 4
+
+#endif
