+++ tiny-firmware/Makefile
@@ -80,6 +80,7 @@ OBJS += protob/c/messages.pb.o
 OBJS += protob/c/types.pb.o
 
 #skycoin-crypto
+OBJS += $(TOP_DIR)vendor/skycoin-crypto/skycoin_signature.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/skycoin_crypto.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/check_digest.o
 OBJS += $(TOP_DIR)vendor/skycoin-crypto/tools/base58.o
