+++ tiny-firmware/firmware/fsm_impl.c
@@ -32,7 +32,9 @@
 #include "recovery.h"
 #include "reset.h"
 #include "rng.h"
+#include "skycoin_constants.h"
 #include "skycoin_crypto.h"
+#include "skycoin_signature.h"
 #include "skyparams.h"
 #include "skywallet.h"
 #include "storage.h"
